gbv_project/requirements.txt
asgiref==3.9.1
attrs==25.3.0
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.3.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
